+++ ByTheBookPlugin.cs
@@ -1,25 +1,21 @@
 ﻿using BepInEx;
 using BepInEx.Logging;
 using BepInEx.Unity.IL2CPP;
-using ByTheBook.SyncDisks;
+using ByTheBook.Dialog;
 using HarmonyLib;
 using Il2CppInterop.Runtime.Injection;
-using System.Collections.Generic;
-using System.Linq;
-using UnityEngine;
 
 namespace ByTheBook
 {
     [BepInPlugin(MyPluginInfo.PLUGIN_GUID, MyPluginInfo.PLUGIN_NAME, MyPluginInfo.PLUGIN_VERSION)]
     public class ByTheBookPlugin : BasePlugin
     {
+        // Normally, I'd advise against using the Singleton pattern so heavily in favor of more decoupled event driven architecture,
+        // but doing so is following the paradigm of the game. Many classes have an 'Instance' singleton in the code as it was designed for 
+        // single player only.
         public static ByTheBookPlugin Instance { get; private set; }
+
         public static ManualLogSource Logger;
-
-        // preset name, preset
-        public Dictionary<string, SyncDiskPreset> byTheBookSyncDisks = new Dictionary<string, SyncDiskPreset>();
-
-        private HashSet<ByTheBookSyncEffects> enabledUpgrades = new HashSet<ByTheBookSyncEffects>();
 
         public override void Load()
         {
@@ -28,14 +24,15 @@ namespace ByTheBook
                 Log.LogInfo($"Plugin {MyPluginInfo.PLUGIN_GUID} is disabled.");
                 return;
             }
+
             Instance = this;
             Logger = Log;
 
             Log.LogInfo($"Plugin {MyPluginInfo.PLUGIN_GUID} is loaded!");
+            
+            RegisterTypes();
 
             PerformHarmonyPatches();
-
-            RegisterTypes();
         }
 
         private void PerformHarmonyPatches()
@@ -47,23 +44,8 @@ namespace ByTheBook
 
         private void RegisterTypes()
         {
-            ClassInjector.RegisterTypeInIl2Cpp<PrivateEyeSyncDiskPreset>();
+            ClassInjector.RegisterTypeInIl2Cpp<ByTheBookDialogManager>();
             Log.LogInfo($"Plugin {MyPluginInfo.PLUGIN_GUID} has added custom types!");
-        }
-
-        public void EnableUpgrade(ByTheBookSyncEffects effect)
-        {
-            enabledUpgrades.Add(effect);
-        }
-
-        public void DisableUpgrade(ByTheBookSyncEffects effect)
-        {
-            enabledUpgrades.Remove(effect);
-        }
-
-        public bool IsUpgradeEnabled(ByTheBookSyncEffects effect) 
-        {
-            return Game.Instance.giveAllUpgrades || enabledUpgrades.Contains(effect);
         }
     }
 }