+++ ModFolderContent/CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.1.0
+
+* Previous functionality moved to the "upgrade 1" slot.
+* Main Effect changed: The Enforcer guard on duty at a crime scene can be asked for a guest pass. 
+  * Higher social credit affects the odds of success up to a 75% success rate.
+  * The "guard" is usually found standing against a wall after crime scene tape has been put up.
+* A configuration option has been added to the BepInEx config to be able to change the price of the PrivateEye sync disk.
+* A configuration option for the SocialCreditLevel for the max success rate of asking a guard for a pass has also been added.
+
 # 0.0.3
 
 * Fixed mod overwriting all other instances of sync disks.